Cleanup patch set (#370)
The [next upgrade](https://github.com/pulumi/pulumi-linode/issues/369)
failed because upstream changed their docs structure. We are compatible
with the new docs structure, but our patches are not.

I'm going to remove ~as much of~ the docs patch set as I can without any
loss in quality before I try to run the update.